Moving auxiliary classes' locations
